Crate fasten_repair

Repairs reads from fasten_inspect output

Examples

./target/debug/fasten_inspect  < testdata/four_reads.fastq | \
  ./target/debug/fasten_repair --remove-info > repaired.fastq

If remove-info is given, then extra header information from fasten_inspect will be removed.

Usage

Usage: fasten_repair [-h] [-n INT] [-p] [--verbose] [--version] [--min-length INT] [--min-quality FLOAT] [--remove-info] [-m STRING]
Options:
   -h, --help          Print this help menu.
   -n, --numcpus INT   Number of CPUs (default: 1)
   -p, --paired-end    The input reads are interleaved paired-end
       --verbose       Print more status messages
       --version       Print the version of Fasten and exit
       --min-length INT
                       Minimum read length allowed
       --min-quality FLOAT
                       Minimum quality allowed
       --remove-info   Remove fasten_inspect headers
   -m, --mode STRING   Either repair or panic. If panic, then the binary will
                       panic when the first issue comes up. Default:repair

Methods of repair

Repair mode

If you choose --mode repair, then this is the expected behavior

Mismatched seq and qual lengths: seq or qual length will be truncated

R1 or R2 not passing: then the other (R1 or R2) will also not pass and they will not be printed.

Panic mode

seq length < min length

avg qual < min qual

invalid characters in seq

invalid characters in qual

@ not present in first character of the entry

+ not present in the first character of the third line

Functions

main 🔒

repair_one_read 🔒

Repair exactly one read

repair_reads 🔒

Repairs reads depending on the deflines by calling repair_one_read